BSIMUangkuBillPayments

+++ Hub/BSMUangkuBillPayments/src/main/java/com/mfino/billpayments/bsm/uangku/BillInquiryProcessor.java
@@ -107,7 +107,7 @@ public class BillInquiryProcessor extends BillPaymentsBaseServiceImpl implements
 		Subscriber subscriber = subscriberMdnService.getSubscriberFromMDN(mdn);
 		String result = null;
 		if (subscriber != null) {
-			result = JSONUtil.getOperatorDescription(responseCode.toString(), subscriber.getLanguage());
+			result = JSONUtil.getOperatorDescription(responseCode.toString(), (int)subscriber.getLanguage());
 		} else {
 			result = JSONUtil.getOperatorDescription(responseCode.toString(), 0);
 		}
